+++ test/anchor_test.dart
@@ -5,12 +5,12 @@ void main() {
   group('Anchor', () {
     test('can parse to and from string', () async {
       expect(Anchor.center.toString(), 'center');
-      expect(Anchor.parse('topRight'), Anchor.topRight);
+      expect(Anchor.valueOf('topRight'), Anchor.topRight);
 
       expect(Anchor.values.length, 9);
 
       for (final value in Anchor.values) {
-        final thereAndBack = Anchor.parse(value.toString());
+        final thereAndBack = Anchor.valueOf(value.toString());
         expect(thereAndBack, value);
       }
     });