+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  - Overhaul the draggables api to fix issues relating to local vs global positions
  - Preventing errors caused by the premature use of size property on game
  - Fix `SpriteAnimationComponent.shouldRemove` use `Component.shouldRemove`
+ - Add utility methods to the Anchor class to make it more "enum like"
 
 ## 1.0.0-rc6
  - Use `Offset` type directly in `JoystickAction.update` calculations